requirements.txt
flask
flask-limiter
flask-cors
google-generativeai
python-dotenv